Refactor TileSpace imports and update references to coordinate_space for consistency
